+++ app/src/main/java/com/appadore/quiz/utils/Constants.kt
@@ -1,0 +1,11 @@
+package com.appadore.quiz.utils
+
+object TimerTime{
+
+    //Time taken to load next question after timeout/answered by user
+    const val QUESTION_TIME_GAP = 5000L
+
+    //Time allocated for a single question
+    const val QUESTION_TOTAL_TIME = 30000L
+
+}
